feat: :sparkles: add ionic plugins
These plugins are: •	@capacitor/camera = provides the ability to take a photo with the camera. •	@capacitor/preferences = a simple key/value persistent store for lightweight data (localStorage). •	@capacitor/filesystem = API for working with files on the device. •	@ionic/pwa-elements = Allows to you us the Camera and other features when not running natively on your mobile device. In file src/main.ts imported the @ionic/pwa-elements plugin.